refactor(shell): drop nushell and bash in favor of ZSH.
nushell is quite innovative, and the main reason for me to switch back
then, was it providing a sane default configuration, without the fuss
of plugins.

However, over time it proved to be challenging, given the unstability
of code base, and introducing breaking changes quite often. Also the
requirement of the compile from source periodically on Nix, made me
wonder if I'm gaining enough value out of the required time for
maintaining it.

In this PR I'm dropping literally any YAGNI I had in my CLI
configuration. The goal is to keep the ZSH setup as pure as possible,
and don't let plugins to effect the performance.