refactor(emacs): only enable Copilot if I want to.
This is in order to reduce its influence on my programming skills. I
can enable it when there are too many chores in the code, or need some
influence, but don't want it to be enabled everywhere.

This is an interesting experiment, maybe I don't really need it that
much.